# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for golang nuryanto2121",
        "title": "Base Playtopia",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/cms/sku-management": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SkuManagement"
                ],
                "summary": "GetList SkuManagement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SkuManagement"
                ],
                "summary": "Create SkuManagement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "this model set from firebase",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddSkuManagement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cms/sku-management/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SkuManagement"
                ],
                "summary": "GetById SkuManagement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SkuManagement"
                ],
                "summary": "Update SkuManagement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "this model set from firebase",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddSkuManagement"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SkuManagement"
                ],
                "summary": "Delete SKuManagement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cms/term-and-conditional": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TermAndConditional"
                ],
                "summary": "GetById TermAndConditional",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TermAndConditional"
                ],
                "summary": "Create TermAndConditional",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "this model set from firebase",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TermAndConditionalForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "errors": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.AddSkuManagement": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "is_bracelet": {
                    "type": "boolean"
                },
                "price_weekday": {
                    "type": "number"
                },
                "price_weekend": {
                    "type": "number"
                },
                "sku_name": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseModelList": {
            "type": "object",
            "properties": {
                "data": {},
                "last_page": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.TermAndConditionalForm": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}